(GH-1845) Add `bolt-defaults.yaml` configuration file
This adds support for a new config file named `bolt-defaults.yaml`. This
file is loaded from both the system-wide and user-level config
directories, and contains similar config options to a `bolt.yaml` file
(except for project-specific config such as `inventoryfile`).

When a user has a `bolt-defaults.yaml` file alongside a `bolt.yaml` file
in the system-wide or user-level directories, it is loaded instead of
the `bolt.yaml` file and warns the user that the `bolt.yaml` file has
been ignored.

If a user has a `bolt.yaml` file present in either the system-wide or
user-level directories, Bolt will issue a warning that the file is
deprecated and to use a `bolt-defaults.yaml` file instead.

!feature

* **Add `bolt-defaults.yaml` configuration file**
  ([#1845](https://github.com/puppetlabs/bolt/issues/1845))

  Bolt now supports a new `bolt-defaults.yaml` configuration file in the
  [system-wide and user-level
  directories](https://puppet.com/docs/bolt/latest/configuring_bolt.html).
  This configuration file is intended to replace the `bolt.yaml`
  configuration file in the system-wide and user-level directories in
  a future version of Bolt. If a `bolt-defaults.yaml` file exists
  alongside a `bolt.yaml` file, Bolt will ignore the `bolt.yaml` file.

!deprecation

* **System-wide and user-level `bolt.yaml` is deprecated**
  ([#1845](https://github.com/puppetlabs/bolt/issues/1845))

  The [system-wide and
  user-level](https://puppet.com/docs/bolt/latest/configuring_bolt.html)
  `bolt.yaml` files have been deprecated in favor of
  `bolt-defaults.yaml`.