(Docs) Bolt commands update
I thought about this some more and I think we're trying to do too
much in this docs: we're giving the syntax for all Bolt commands, but
also trying to explain how to interact with remote targets. Based on
the amount of content that's aimed at interacting with targets, I think
a user is actually coming here to figure out how to interact with
targets using Bolt commands, and not for help running commands in
general. With that in mind, I've moved the syntax section to the Bolt
command reference template (And removed the powershell syntax bit for
now).

I've also tried to flatten the structure of the doc a bit more and I've
built in H3s to make it a little easier to skim. This has allowed me to
remove the bullets from the command blocks. The page was looking really
busy with all of the bullet points and indentation.